teacher_server/requirements.txt
flask
flask-socketio
websockets
pyjwt
pywin32
pyautogui
pillow
pyinstaller